Closes #1275, Map RIS ED tag to editor
Mapping both import and export. The spec doesn't actually give any indication that ED is for editors (and, given that AB isn't actually abstract, you can't really make assumptions when it comes to RIS), but other apps seem to put editors in ED

A forum user suggested that mixing and matching A1/A2 and ED caused creator data to be skipped altogether on import to EndNote. I haven't checked this, but I'm adding editor export to ED on the grounds that more data is better than less, and users can update other import filters if necessary.