Don't use API in Amazon.com translator
Amazon is going to revoke our API key today because we are not allowed
to make it publicly accessible. We don't see a good way around this.
This translator looks up ISBNs at Open WorldCat (translator ID
hard-coded, since the current connector doesn't deal well with search
translator failure), or else tries to scrape the page if there is no
ISBN.

It's not going to work as consistently as the old translator, and it
will need a lot of work to be able to scrape pages without ISBNs in
non-English languages, but it's better than nothing.