Added collaborators and checking the primary creator type
I've also added some commented code in the generation of cite keys. Please consider this for integration with a Zotero version that allows accessing configuration parameters from Zotero.Utilities.

The way creators are sorted does not consider whether they are primary authors or not, when adding some cite to some reference this is puzzling, because the key can be generated from a surname that isn't that much relevant, and it's also hard to remember.

This could not be configurable (nor commented) and forced to be so, but many people could be upset if their citekeys stop working, getting LaTeX warnings and whatnot.