[DOI] Add search component to the DOI translator The search translator serves as a dispatcher for other DOI translators: 1. First, it uses CrossRef to look for a DOI, since most DOIs we deal with are from CrossRef 2. If not found, it queries CrossRef's RA lookup tool to determine which RA is responsible for the DOI 3. If available, it uses the appropriate DOI search translator to query the DOI
DOI search translator takes a string (corresponding to a DOI), an item containing a DOI property, or an array containing a combination of these.

The search translator will call `item.complete()` for each DOI queried. `item._query` will be set to the original query corresponding to the resulting item. `item._status` will be a string corresponding to the status of the query. If successful, the status will be 'success'. If the search fails for one of a number of reasons, the status will indicate the reason. Some possible values for `item._status` are 'not found', 'invalid', 'unimplemented', and 'fail'. These are not set in stone.

DOI search translator will pass an array containing the original queries (filtered depending on RA) to the downstream DOI translators. The queries will contain valid DOIs (except those passed to CrossRef) either as literal strings or objects containing DOI properties (or a combination thereof). However, DOI search translators should not assume queries to be valid, because they can be invoked directly from other translators.

DOI search translator should server as the only DOI translator responding to `detectSearch` for DOI queries. All other DOI search translators should not return `true` for these queries. This behavior is overridden (inverted) by setting `directSearch` property to `true` on the query object.

These changes depend on the ability to pass objects associated with items between translators. Implemented in zotero/zotero#401