Update Primo.js: Creators are now saved as lastName, firstName
This fix will try to better fetch the creators for Primo.js, such that they will be splitted into lastName and firstName as asked in https://forums.zotero.org/discussion/43372/individual-book-authors-display-as-corporate-authors-in-ex-libris-primo/ . The approach will leave real corpoorate authors as they are, but for real persons tries to fetch the same person in a normalized form (lastName, firstName) from the addata section, if possible. I tested it with several primo instances and it seems to work well.