Work for new Google Patents format;
This works nicely for individual articles, but I simply can't make it happen for multiples with all of the JS templates that build up and change the different parts of the site. I'd be happy for more input from @aurimasv or @zuphilip or we can merge this since it does work nicely on individual articles.

Tests don't work either.