Update Retsinformation.js
I have added test cases. And I've made a few changes:

Documents from "Folketingets Ombudsmand" (an independent unit under the Danish Parliament) has been changed from 'report'-type to 'webpage', as it seems more appropiate.

The HTML used for bills is very varied, sometimes including "pretitle"-elements, som of which contain part of the proper title, and some containing other information. So I chose to use the head-element to get titles for all bills. I understand that getting the title from the head-element is discouraged, but it seems to me, to be the best solution in this case.

- regards
Roald