Update mirrors handled by the Theory of Computing translator
The mirror for Theory of Computing at the Anustralian National University
is now offline, while a new mirror has come up at the New York University.
This commit updates the regex and tests for the transltor to reflect this
change.

The commit also fixes a small typo in one of the tests.