[don't merge] Switch DOI to DOI.org
@zuphilip -- this obviously needs cleanup, but I wanted to get your
thoughts on this. It's taking advantage of the now available CSL_JSON
via content negotiation on doi.org (and the functionality you added to
ZU.doGet to use it. The main advantages are
1) Fewer API requests
2) Works in Chrome connector for items with DataCite and CrossRef DOIs
3) Will work with additional services as they become available

Open questions:
1) What's a good name for the new translator
2) Should we remove the old DataCite and CrossRef ones, or leave them in
as a fallback at least initially?
3) Any thoughts on how to test this for data quality? I was going to
manually go through all DOIs that are in the CrossRef translator. Any
other ideas?

As well as any other conceptual issues you can think of (as I say, don't
worry about code quality for now)