Update Library Catalog (PICA).js (#1293)
ABES has been updating the cataloguing recommandations to FRBR/RDA and on some records, the display has been updated with new labels like : http://www.sudoc.fr/200278649 (Publication instead of Editeur(s) for example). The purpose of this patch is to fix the issue that happens actually which is that publisher information are not found by Zotero for these newly created/updated records.