test for pdfurl
Thanks for clarifying that.

But if the translator being unable to find an url could be an error, there is another issue that I kind of ignored.
The website will sometimes give just the name of the journal as a search result. Zotero doesn't have an item type for an entire journal, so I just ignored it, but there is a possibility that one can select a journal name from the select screen. I actually did it, and Zotero just ignored it. Should I do something about this?