Fix ScienceDirect PDF downloads (#1381)
On many ScienceDirect pages, it's now necessary to load an intermediate
page with a meta refresh to the final CDN URL for the PDF, and to get the
intermediate page URL we have to either simulate a click to open a menu
or load the unprocessed HTML source of the page. Hideous, but it works.

This also uses a new method for determining whether the user has access
to the PDF or would get a "Your organization may be charged for
transactional access to this non-subscribed article." page (which has a
Continue button, but presumably we don't want to touch that).