[BibTeX] Use location as event-place for conferences
Moreover, in this way we make sure that the order of the fields
`address` and `location` in BibTeX don't matter. We take `location`
either as `event-place` in extra for conferenceArticles (e.g. for
ACM) or as a backup for the `place` field. This solves #1020.

Additionally, delete `institution` from inputField map because it
is already handled as an backupPublisher.