[EM] Don't concatenate different nodes by default (#1595)
* [EM] Don't concatenate different nodes by default

This lead to title and abstracts containing also a
translated form, see examples in Cairn.info.js.

* Update tests, change handling of og:site

- The publicationTitle is also considered for cleaning title
when this information comes from doc.title. However, after
e7493f3 this information was not always given, e.g. test case
http://volokh.com/2013/12/22/northwestern-cant-quit-asa-boycott-member/
Therefore, we switch here the condition but left the field the same.
- DIVA is the site name in og:site and not the proceeding title.
- The salon website has moved most of its meta tags into the body
and therfore the recognition is quite bad.
- The detection of Hindawi test case still fails, but maybe the can
fix this...

* explain `allValues` variable in comment