TEI: Use Better BibTeX citation key if available (#1810)
This will use the citation key generated by Better BibTeX for Zotero as the `xml:id` if it is available. If that add-on is not enabled, this change has no effect. Solution from @retorquere in <https://github.com/retorquere/zotero-better-bibtex/issues/1124>.

Note discussion on PR at https://github.com/zotero/translators/pull/1810