Fixing date, abstract and multiple in YouTube.js
This fixes https://forums.zotero.org/discussion/78171/zotero-does-not-detect-date-of-youtube-video

There was no other data about the date found on
the youtube page itself, but strToISO seems to handle
these cases currently quite well.

The test cases in Scaffold will only work
with waiting for some time for the pages to load.
Thus, the defer option was added, but currently this
only works for updating the tests.