Fix PDF url in ACS Publications.js and clean up
The xpath for the PDF url (itemOpts.pdf) did match with
two nodes and therefore outputting a concatenated non-valid
string for it. This is fixed by focusing on the first match.

Additionally try to enable PDF saving in multiples directly.

Moreover, delete the different parts for High-Res PDF resp
PDF Plus with Links, which does not exist anymore.

This solves #1966 now completely.