WorldCat Discovery Service: Throw error on non-RIS response
For me, both WorldCat Discovery Service and Open WorldCat detect on this
page [1], but a 404 is returned for citation.ris for WDS and OW runs
instead. Someone in the forums [2] is getting a 200 response code for a
404 error page ("Leider nichts gefunden. Sie können in
WorldCat-Bibliotheken fast alles finden, aber leider nicht an dieser
Stelle."), which was causing WDS to complete without error but without
saving anything. This changes the WDS translator to search for the "TY
- " RIS header in the response and fail explicitly. It would probably be
better for the RIS sub-translator to just reject if no RIS was found,
but that might need to wait for proper async translator support.

I'm also not sure why both WorldCat Discovery Service and Open WorldCat
are triggering on the same URL. Best case, this will show a fallback
error in the save popup. It seems like it'd be better to combine these
into a single WorldCat translator.

[1] https://www.worldcat.org/title/righteous-mind-why-good-people-are-divided-by-politics-and-religion/oclc/897080656&referer=brief_results
[2] https://forums.zotero.org/discussion/79621/worldcat