Fix detectWeb and add ISBN to Internet Archive translator
* Books on IA have ISBNs but this info isn't being
scraped. Add support to add this info and add
book to test with multiple ISBNs.

* Fix xpath for detectWeb / icon
Upstream, the icon that indicates the
itemType is no longer within the h1 tag,
so detectWeb was broken. This fixes it
so the xpath is now able to select the div
containing the icon and use this to
determine item type.