APA PsycNet: Trigger on additional URL patterns
This stupid site blocks on a hair trigger, which makes it impossible to
work with, so I can't actually test if this works, but this at least
makes the icon appear correctly on more pages.

I'm not sure why /buy/ wasn't included in detectWeb() before, since
there's saving code that seems to handle it.