CSL JSON: Default to 'article' if no 'type'
'type' is required in CSL-JSON, but some DOI registration agencies
provide bad data, and this is better than failing. (itemFromCSLJSON()
will already default to 'article' for unknown 'type' values.)

Technically this should go in the DOI Content Negotation translator, but
it's easier to do this here after the JSON has been parsed, and it might
benefit other translators.

This is just for imports from other translators. File/clipboard imports
without 'type' still won't work, because a valid 'type' is required in
detectImport(). (This also can't be tested in Scaffold, which does a
detect first.)

https://forums.zotero.org/discussion/85273/error-importing-dois-via-add-item-by-identifier