CI: Check deleted.txt and lint even if earlier steps fail (#2389)
Unless your `if:` expression contains either `success()` or `failure()`,
GitHub will default to running only on success. We probably can't run
tests if the setup steps fail, but we can usually still lint and check
deleted.txt even if tests fail. It's nice to get as much information as
possible in one CI run when making a PR.