YouTube: Scrape correct creator and description
The YouTube translator was sometimes grabbing the wrong channel name.
How is this possible when we're selecting by ID? Well,
document.querySelectorAll('#text-container') returns 22 nodes.

Whatever.

The other issue was with the description, which started pulling
surrounding UI elements and metadata.

Whatever the case, these selectors are a bit better, and it'll still
fall back on the less specific originals.