Retsinformation: Use JSON, not XML, for same results as old translator
The XML was nice, but it's slow to download and not focused on providing
human-readable data (which is what we want for most fields). The JSON
gives us the date in a form that's a bit *too* human readable, but
strToISO can parse it. Making two requests would be the best of both
worlds, but I don't want to grab the XML if it isn't strictly necessary,
and it's not.