Trove: Handle books, articles, art, and multiples
We lost tags on works - they're difficult to clean and usually a mix of
LoC subject headings, restatements of the title, date ranges, and
various permutations of the authors' names. Just not very meaningful.

We emulate a browser and use the same API key as the site's JS in order
to grab the BibTeX. I'd like to find a better approach. I don't
understand why citation generation is locked behind an API key wall.