NASA ADS: Generate token for new API
ADS has a nice new API, but it requires a token to access. I
experimented with creating an account and writing a proxy, but I
realized that there's a much simpler way: we can just generate a
temporary token exactly as the site does when you generate a citation in
the browser, by calling the '/accounts/bootstrap' endpoint. Works great.

Updated old tests and added a new one. Everything now seems to be
working just as it was before the update.