Add translator for Web of Science Nextgen (#2464)
Web of Science is running two sites in parallel right now: a legacy site
and a beta "Nextgen" or "New" site. The "New" one is set to become the
default soon, but the existing WoS translator doesn't support it at all.
It now uses Angular and doesn't easily give us any of the form data that
the existing translator uses in order to pull item metadata. I didn't
want to clog up the old translator with layers and layers of
conditionals, so I created a new one.

The least dependable part right now is the URL. The translator tries to
pull it from the "Full text at publisher" button on the page, but that
doesn't show up in the static HTML, so search results don't get URLs.
I'm not sure what to do about that yet.

The app itself populates its model via RPC over a WebSocket.