WorldCat Discovery: Support v2 site
I pursued two metadata sources for this: RIS, which the original
translator uses, and internal WorldCat metadata. WorldCat metadata is
promising: it includes correct creator types (everyone is AU in the
RIS), cleaner creator names, and an abstract. But it's locked behind a
client-generated API key system that requires access to the JS HMAC API.
I wrote a hosted lambda function to generate the required values, but
after a day of working on a translator, I was still only getting
passable data. It was a fragile approach with significant data-cleaning
issues.

Although there are downsides to the RIS they give us, users are already
used to them. We can stick with it for now.