ScienceDirect: Don't detect search pages as journal articles
detectWeb uses journalArticle as a fallback type after tests for all
other types have failed (the opposite of the way most translators do it,
but it makes sense for this site). That was causing empty/not-yet-loaded
search pages to show as journalArticles, and the toolbar icon wouldn't
update once loading completed.

Also removed attr/text and added a license block.