METS: Throw an error if metadata type is unknown but not 'OTHER'
'OTHER' usually indicates internal metadata, and it can be of any type.
In one of the test cases, it's used to embed an internal image metadata
format. There's no way that we can ever make sense of it, but it is very
unlikely to be essential to importing the item fully, so we'll just skip
it. In all other unknown metadata cases, abort the import.