Crossref Unixref XML: Bump timestamp
We've had a couple reports of this file being missing [1], causing PDF
metadata retrieval to freeze. I'm not sure how that would've happened,
but it's possible it's more fallout from the renamed-translator bug that
was fixed in bde9a74f9, which still exists on 5.0.96.2.

To be safe, push this out again, which should cause it to be recreated.

[1] https://forums.zotero.org/discussion/90750/progress-window-stuck-during-pdf-metadata-retrieval