Twitter: Don't offer to translate when regex will fail
This primarily affects when a user opens a tweet by clicking on the
image, not the text, to get the full-screen lightbox view. Twitter
doesn't update the page title in that case, so it's very difficult for
us to scrape anything useful. Previously the translator would detect
'blogPost' but error out in `doWeb`; now we'll just not detect anything
at all. Better than a confusing error.