beck-online: Allow detection of proxied URLs
This is a hack. EZproxy replaces dots with dashes in proxied HTTPS URLs,
so the connector undoes that by replacing dashes with dots.
Unfortunately, that means that beck-online.beck.de will be proxied to
beck-online-beck-de and then de-proxied by the connector to
beck.online.beck.de. Accepting a dot in addition to a dash sidesteps the
issue.