ADS: Support arXiv preprints, save citation key
ADS can work as an odd kind of proxy for arXiv when a bibcode starts
with the publication year followed by `arXiv`, so we should support
that.

Saving the citation key is mostly helpful for BBT users with an
ADS-heavy workflow who use bibcodes as citekeys (which seems to be
common), but it's worth having.