Add translator for LoC.gov entries
The Library of Congress has a few catalogs, but this is the main search
portal for all of them. We bypass the Voyager 7 translator for
lccn.loc.gov because of some problems running it within
processDocuments, and we try to filter out direct links to non-catalog
pages from search results.

Results won't actually be as nice as the tests until the MARC PR (#2576)
is merged, specifically because of some incorrect logic that changed
artworks to journalArticles.