SSRN: Update multiples selectors, lint
Instead of detecting multiples via URL, we now use the now-standard
approach, and we're able to support more multiples pages (author
listings, journal pages, top articles lists, etc.). I'm not sure if the
two XPath-based multiples selectors are ever actually used now, but I'm
fine with leaving them for now just in case. SSRN has a lot of different
structures for search result and journal listing pages, so it's better
to be safe.