Add translator for CQ Press publications (#2683)
I'm unsatisfied with a few aspects of this translator. The RIS they give
isn't excellent, but in some cases there's nothing we can do (e.g., in
the "Bicknell John" case, a human can tell that that's clearly a
reverse-order name, but the translator can't). There's a great diversity
of item types and search result formats on the site and I'm not
confident that I've handled every case. But this is a good start, at
least, and gives us full coverage with some imperfections.