WorldCat Discovery: Support non-English interfaces
Reported:
https://forums.zotero.org/discussion/comment/367201/#Comment_367201

This prioritizes a successful COinS-based type detection over the UI
scraping-based method we were using before. We sometimes get incorrect
type detections on fresh page loads, but the worst case is an article
being detected as a book, and that's better than no detection at
all. (data-testid turns out to have been as unstable as the comment
predicted it would be.)

Tests are a mess right now - v1 tests no longer work because the sites
they target have updated to v2. That'll wait for a separate commit.