new translator for online journal the PRC history review (http://prchistory.org/the-prc-history-review/)
The site uses wordpress with direct file link without dedicated child page for individual articles for most cases, so I use less structured ways to extract the information.

It support downloading articles from the home page of the Review (http://prchistory.org/the-prc-history-review/). However, since this home page has entries that are ever-growing, I did not include it in the test.