K10plus ISBN: Force PDF ToC URLs to HTTPS
www.gbv.de and d-nb.info support HTTPS, and I'm guessing any other
domains specified in MARC records do too. This is necessary for PDF
downloads to work for Add by Identifier on iOS and is just generally a
good thing to do. If we find specific domains that don't work we can add
exceptions for those.