CSV: Fix regressions after 21819b8bbd
Revert to legacy mode, which includes base-mapped fields in
`uniqueFields`. In non-legacy mode, we don't seem to have any access to
base field mappings (hence the bad hard-coded mappings in, e.g., EndNote
XML). We'll need to add a way to get base fields for the export
translators that need it.

Also:

- Fix how tag types are handled so tags don't break when we do update
  from legacy mode

Fixes #2843