ePrint: Retrieve author names from meta tags
The extraction of authors from the web page was broken.  In particular
when authors are formatted as "First1 Last1, First2 Last2, and First3
Last3", then splitting by "and" results in the names of the first two
authors getting mixed up.

The HTML meta tags seem to be a more reliable source for the authors.