Add Korean National Library
Library has two typs of resources, one comes in MARC, one comes in MODS; the MODS also requires cleaning up item types.

Given how item types in MODS are set/determinde, I decided to swap out the text in the downloaded MODS rather than parsing the XML and mapping it. It's a little harder to read this way, but takes advantage of the code in MODS that determines item types properly.

Tests time out, but I thought it'd be useful to have them (the URLs are stable)