Change FAO publication target URL
FAO has changed the URL naming scheme. Now fao.org/publications/* and fao.org/documents/* are not longer the same, and this script only applies to the former. I'll upload a script that works for the latter soon.

It's likely that FAO will use the "documents" URL as the main link for publications from now on.