Infer the names of podcast guests using the body text.
The body text paragraphs mark the contributor names (including host and
guest) in bold (<b> or <strong>), among others (such as the initials of
their names and some promotional text).

We take these marked fragments, perform suitable normalization (remove
punctuations, etc.), and filter out the initials (all-caps).

For the remaining fragments, to reduce false-positives, we check each
against the title, and only keep the ones that appear in the title.

The final list of names is used as the names of guests.