NASA ADS: Handle the ampersand character (&) in BibCode.
The BibCode (bibliographic identifier) *may* indeed contain the
ampersand character. For example, the journal "Astronomy and
Astrophysics" becomes "A&A" in the BibCode.

The ampersand as part of the URL is always encoded as '%26' (even
decodeURI() will not un-quote it). However, this percent-encoded form
cannot be part of the BibCode. If it is not further decoded back into
the ampersand, the regex for BibCode will not match, which breaks the
translator from the start.

To fix this, the BibCode-extraction routine now performs an extra
decodeURIComponent() for it.

Other mandatory-encoded characters (such as the question-mark and
equal-sign) do not appear in BibCodes.