Novaya Gazeta: initial commit.
The Novaya Gazeta ("New Gazette") is an award-winning newspaper renowned
for its independence.

The NG in its original form ceased operations in March 2022, and its
successor, Novaya Gazeta Europe, started publishing online in April
2022. Articles are available in Russian and English.

Please note that this translator is for the Novaya Gazeta *Europe*
on-line publication, not the original Novaya Gazeta that was shut down.

This translator gathers the article data by querying the NGE's JSON API,
just like the browser. This is necessary, because the browser-rendered
document itself is highly obfuscated.