CI: When no translator is updated in PR, do self-testing.
In 'Test pull request', when the PR is not concerned with a translator,
run the pull-request test with a minimal mock translator. The purpose is
to check that the test itself is functioning. (When no translator is
updated, it is likely that the PR is intended to modify the testing
itself.)