Lapham's: Better de-duplication of titles in search results.
As an accessibility measure, when the search page returns duplicate
titles, they are disambiguated by a number tag.

Previously, the first appearance of a duplicate was not marked by "1",
which is fixed by this commit. Unique titles remain unmarked.

In addition, the URL paths are appended after the number as an extra
assistance.

Change in visual effect in the item selector:

Before:

  - Title
  - Title (2)

After:

  - Title (1, URL: url1)
  - Title (2, URL: url2)