DOI Web translator: Add a test case of DOI with special characters.
We want a form of the online resource that can lead to a
"journalArticle" for testing purposes. As a hack, we use the citation
export page of Wiley Online Library, rather than the journal article's
own "canonical" page, because it

- has the (urlencoded) DOI of the item in the URL;
- does not contain any links with DOI for *other* resources such as
  bibliography or also-read articles; and
- loads very fast.

The expected behaviour is to find one item to save (i.e. not
"multiple"), and obtain it by DOI.

This DOI in particular also contain "special" characters such as
< > : ; etc.