DOI Web translator: Exclude quotation marks from textual DOI pattern.
In the regex meant for scraping the text, don't include quotation marks
anymore.

The regex is also broken into parts so comments can be put next to each
component. We make use of `String.raw` syntax to make the full regex
easier to understand.

A test case is also added to cover the change.