DOI Web translator: Add a testcase with DOI containing hash character #.
This is a particularly interesting case, with the DOI

10.1002/(SICI)1521-3951(199911)216:1<135::AID-PSSB135>3.0.CO;2-#

Ideally, we should detect only one DOI (the correct one) and detect the
type as `"journalArticle"`. However, due to a bug on this website, we
detect another, malformed "DOI":

10.1002/(SICI)1521-3951(199911)216:1<135::AID-PSSB135>3.0.CO;2-

(note the absence of the hash at the end)

The source of this malformed DOI is, ironically, the <a> tag on the DOI
display, under the "How to cite" heading. The text content of the <a> is
the correct DOI, while the href value is malformed (unencoded). Clicking
on the link (correctly) leads to a resolution failure from the doi.org
proxy.

As a result, we detect this as "multiple", although the item selector
only shows the correct entry.