Bosworth-Toller dictionary: New translator initial commit.
The "Bosworth-Toller" is one of the most comprehensive lexicon for the
Old English language (c. 8--11th century). The website
https://bosworthtoller.com/ is a modern digitization of the
Bosworth-Toller dictionary, hosted by the Faculty of Arts, Charles
University, Czech Republic.

This translator works for individual words' pages or the search results.
In addition to the page snapshot, when possible it also automatically
saves a note containing a concise textual summary of the entry's
definitions and grammar info.