Bosworth-Toller dictionary: Correctly identify original book and page.
The entry may originate from either the main dictionary book (1898) or
the supplement (1921). The distinction is correctly applied when
obtaining the original starting-page info.

In addition, the "pages" field is removed (because the online dictionary
does not have page numbers). The page in the original book is now an
extra field.