Google Scholar: Add delays between consecutive network requests.
When there are multiple items to be translated, the previous
implementation fires network requests to Google Scholar server in rapid
succession. This increases the likelihood of being rate-limited or
denied by the server.

To fix this, an artificial delay of at least 500ms is added between
successive network requests.

In addition, calls to the callback-based `doGet()` are replaced with
promise-based asynchronous calls.

A few stylistic fixes (such as `===` vs. `==`) are also included.

Resolves #3037.