Google Scholar: Use external services (DOI and ArXiv API) when possible.
To reduce the load of Google Scholar servers and therefore the
likelihood of getting an error or captcha, we use the DOI and arXiv
identifiers (inferred from article title links) with the corresponding
search translators when possible.

The Google Scholar "native" scrape is used as fallback when the external
search fails.

The requests to different services are done concurrently, but for each
service, requests are sequentialized, and consecutive requests are
interspersed with delays.

If there are items left untranslated at the end of this process, an
error is raised.