zbMATH: Prefer rendered math text in inline contexts (title, tags).
- In titles, the math is typically brief inline text. By using rendered
  text instead of LaTeX source, the saved PDF's file names will contain
  less "special" characters. This improves interoperability, and reduces
  the likelihood of certain user errors with "special" characters in
  file names. In addition, the look and feel will be closer to the
  normal expectation.
- In tags, already brief, the LaTeX math is a distraction.

Note that the MathML extraction routine is tested against most typical
MathJax preferences (and with the Firefox "Native MML" extension). If
Assistive MathML is turned off (default on, and strongly recommended),
the result will be more accurate. In the extreme case of Assistive
MathML off and SVG rendering on, the math text may disappear altogether.