[WIP] WoS Tagged: Reimplement the core algorithms.
The previous iteration of the import translator for Web of Science
Tagged Format is becoming a bit difficult to maintain because of the
tight coupling of logic, data, and actions.

To improve the separation of concerns, in the updated version several
measures have been taken:

- Use separate methods for distinct functions such as line
  parsing/validation (purely formal text-processing), intermediate
  transformations (data normalization based on semantics, etc.), and
  data-to-item mapping.
- Use a lookup-table approach to help with mapping tagged data to object
  properties, which is easier to debug/maintain. It can also better
  support "polymorphism", the unfortunate fact that the same tag can
  mean different things depending on the item type.
- Implement better text-processing (e.g. collapsing spaces when
  necessary, cleaning up line noise, more robust handling of author
  names, etc.)

Overall the goal is to make the translator more robust and easier to
reason with.

TODO: To ensure best compatibility of behaviour, none of the test cases
has been updated for now. After this commit, I'll introduce temporary
devices to ensure the new code, with its new underlying structure,
produces the same output as the old one (to the point of bug-compatible
except for the most egregious). When that is achieved, the temporary
compatibility devices will be removed, and the test cases will be
updated and manually verified. After that, the new test cases will
serve as the basis for incremental improvements of the new code.