EM: Workaround for <meta> from head ending up in body due to invalid HTML
A fairly common consequence of being served invalid HTML is that after
parsing, <meta> elements that should have gone into the <head> end up in
the <body> in the output of DOM parser.

This breaks the EM translator which only looks for <meta> tags that are
immediate descendants of <head> -- the sensible behaviour for
well-behaving documents.

To work around this issue, the default behaviour is not changed, and a
configurable flag is introduced to override it.

When the `searchForMetaTagsInBody` property is set to true, <meta> tags
will be searched for in the immediate descendants of <body>.

With this workaround, Web translators that calls the EM translator don't
have to manipulate the document by putting <meta> elements into the
head manually.

This reduces the likelihood of unintentionally mutating the DOM, and it
keeps the EM-based translator code's cleaner.

A real-world example is the Brill translator, also fixed in this commit.